# server/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "易宿酒店预订平台 API",
    "version": "1.0.0",
    "description": "基于 Node.js Express 的酒店预订系统后端接口文档"
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "本地开发服务器"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Hotel": {
        "type": "object",
        "required": ["merchantId", "name", "city", "address", "starRating", "price"],
        "properties": {
          "merchantId": { "type": "string", "description": "关联商户ID", "example": "609c1..." },
          "name": { "type": "string", "example": "上海中心J酒店" },
          "nameEn": { "type": "string", "example": "J Hotel Shanghai Tower" },
          "city": { "type": "string", "example": "上海" },
          "address": { "type": "string", "example": "上海市浦东新区东泰路126号" },
          "starRating": { "type": "integer", "minimum": 1, "maximum": 5, "example": 5 },
          "price": { "type": "number", "description": "起始价格", "example": 9999 },
          "openingTime": { "type": "string", "example": "2021年" },
          "description": { "type": "string", "example": "云端体验" },
          "tags": { "type": "array", "items": { "type": "string" }, "example": ["地标", "奢华"] },
          "images": { "type": "array", "items": { "type": "string" }, "example": ["/uploads/hotel1.jpg"] },
          "status": {
            "type": "integer",
            "enum": [0, 1, 2, 3],
            "description": "0:待审核, 1:已发布, 2:拒绝, 3:下线",
            "default": 0
          },
          "rejectReason": { "type": "string", "example": "图片不清晰" }
        }
      },
...
      "Order": {
        "type": "object",
        "required": ["userId", "hotelId", "roomTypeId", "checkInDate", "checkOutDate", "totalPrice"],
        "properties": {
          "userId": { "type": "string", "description": "下单用户ID", "example": "user_123" },
          "hotelId": { "type": "string", "example": "hotel_456" },
          "roomTypeId": { "type": "string", "example": "room_789" },
          "checkInDate": { "type": "string", "format": "date", "example": "2023-10-01" },
          "checkOutDate": { "type": "string", "format": "date", "example": "2023-10-03" },
          "quantity": { "type": "integer", "default": 1, "example": 1 },
          "totalPrice": { "type": "number", "example": 2000 },
          "status": {
            "type": "string",
            "enum": ["pending", "paid", "completed", "cancelled"],
            "default": "pending",
            "example": "paid"
          },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      }
    }
  },
  "paths": {
...
    "/api/hotels": {
      "post": {
        "summary": "发布新酒店 (仅商户)",
        "tags": ["Hotels"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Hotel" }
            }
          }
        },
        "responses": {
          "200": { "description": "发布成功" },
          "403": { "description": "权限不足" }
        }
      },
      "get": {
        "summary": "首页搜索酒店 (公共)",
        "tags": ["Hotels"],
        "parameters": [
          { "name": "city", "in": "query", "schema": { "type": "string" }, "description": "城市筛选 (如: 上海)" },
          { "name": "keyword", "in": "query", "schema": { "type": "string" }, "description": "名称或地址关键字" },
          { "name": "starRating", "in": "query", "schema": { "type": "integer" }, "description": "星级 (1-5)" },
          { "name": "minPrice", "in": "query", "schema": { "type": "number" }, "description": "最低价" },
          { "name": "maxPrice", "in": "query", "schema": { "type": "number" }, "description": "最高价" },
          {
            "name": "sortType", "in": "query",
            "schema": { "type": "string", "enum": ["default", "price_asc", "price_desc", "score_desc"] },
            "description": "排序方式"
          },
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 }, "description": "页码" },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 10 }, "description": "每页条数" }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": { "type": "array", "items": { "$ref": "#/components/schemas/Hotel" } },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "total": { "type": "integer" },
                        "page": { "type": "integer" },
                        "limit": { "type": "integer" },
                        "totalPages": { "type": "integer" }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/hotels/my": {
      "get": {
        "summary": "获取我的酒店列表 (仅商户)",
        "tags": ["Hotels"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Hotel" } }
              }
            }
          }
        }
      }
    },
    "/api/hotels/admin/list": {
      "get": {
        "summary": "获取所有酒店列表 (仅管理员)",
        "tags": ["Hotels"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Hotel" } }
              }
            }
          }
        }
      }
    },
    "/api/hotels/{id}": {
      "get": {
        "summary": "获取酒店详情",
        "tags": ["Hotels"],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Hotel" } }
            }
          },
          "404": { "description": "未找到酒店" }
        }
      },
      "put": {
        "summary": "修改酒店信息 (仅商户)",
        "tags": ["Hotels"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/Hotel" } }
          }
        },
        "responses": {
          "200": { "description": "修改成功" }
        }
      }
    },
    "/api/hotels/{id}/audit": {
      "put": {
        "summary": "审核酒店 (仅管理员)",
        "tags": ["Hotels"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": { "type": "integer", "description": "1:通过, 2:拒绝" },
                  "rejectReason": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "操作成功" }
        }
      }
    },
    "/api/hotels/{id}/status": {
      "put": {
        "summary": "酒店上下线 (商户/管理员)",
        "tags": ["Hotels"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": { "type": "integer", "description": "1:上线, 3:下线" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "操作成功" }
        }
      }
    },
...
    "/api/orders": {
      "post": {
        "summary": "创建订单",
        "tags": ["Orders"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["hotelId", "roomTypeId", "checkInDate", "checkOutDate"],
                "properties": {
                  "hotelId": { "type": "string" },
                  "roomTypeId": { "type": "string" },
                  "checkInDate": { "type": "string", "format": "date" },
                  "checkOutDate": { "type": "string", "format": "date" },
                  "quantity": { "type": "integer" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "下单成功",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Order" } }
            }
          },
          "400": { "description": "参数错误" }
        }
      }
    },
    "/api/orders/my": {
      "get": {
        "summary": "获取我的订单列表",
        "tags": ["Orders"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Order" } }
              }
            }
          }
        }
      }
    },
    "/api/orders/{id}/cancel": {
      "put": {
        "summary": "取消订单",
        "tags": ["Orders"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "取消成功",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Order" } }
            }
          }
        }
      }
    },
    "/api/favorites": {
      "get": {
        "summary": "获取我的收藏列表",
        "tags": ["Favorites"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "获取成功" }
        }
      }
    },
    "/api/favorites/{hotelId}": {
      "post": {
        "summary": "收藏酒店",
        "tags": ["Favorites"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "hotelId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "201": { "description": "收藏成功" },
          "400": { "description": "已收藏过" }
        }
      },
      "delete": {
        "summary": "取消收藏",
        "tags": ["Favorites"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "hotelId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "取消成功" }
        }
      }
    },
    "/api/favorites/check/{hotelId}": {
      "get": {
        "summary": "检查是否已收藏 (用于前端UI状态)",
        "tags": ["Favorites"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "hotelId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "isFavorite": { "type": "boolean", "example": true }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}